separate out header filter gin handlers, start adding swagger docs
