+++ internal/api/client/admin/headerfilter_get.go
@@ -1,0 +1,36 @@
+// GoToSocial
+// Copyright (C) GoToSocial Authors admin@gotosocial.org
+// SPDX-License-Identifier: AGPL-3.0-or-later
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package admin
+
+import "github.com/gin-gonic/gin"
+
+func (m *Module) HeaderFilterAllowGET(c *gin.Context) {
+	m.getHeaderFilter(c, m.processor.Admin().GetAllowHeaderFilter)
+}
+
+func (m *Module) HeaderFilterBlockGET(c *gin.Context) {
+	m.getHeaderFilter(c, m.processor.Admin().GetBlockHeaderFilter)
+}
+
+func (m *Module) HeaderFilterAllowsGET(c *gin.Context) {
+	m.getHeaderFilters(c, m.processor.Admin().GetAllowHeaderFilters)
+}
+
+func (m *Module) HeaderFilterBlocksGET(c *gin.Context) {
+	m.getHeaderFilters(c, m.processor.Admin().GetBlockHeaderFilters)
+}
